requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
Bottleneck==1.4.2
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.1
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
executing==0.8.3
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.11.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
jiter==0.9.0
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.6
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
nest_asyncio==1.6.0
numexpr==2.10.1
numpy==2.2.4
openai==1.72.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pip==25.0
platformdirs==4.3.7
prompt_toolkit==3.0.43
proto-plus==1.26.1
protobuf==5.29.4
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.15.1
pymongo==4.12.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
pywin32==308
pyzmq==26.2.0
requests==2.32.3
rsa==4.9
selenium==4.31.0
setuptools==72.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.2.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
typing-inspection==0.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.5
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
